refactor : status 변경 기능 ResponseDto 파라미터 변경
- Member가 아닌 MemberDto로 나중에 변경할 예정인데 Member 구현이 머지되면 Dto 생성 후 붙일 예정이라 따로 분리했습니다.
